Refactor: refactor mypage 마이페이지 폴더로 나누어 리팩토링
