+++ alcoholmap/src/frontend/pages/mypage/index.tsx
@@ -1,4 +1,4 @@
-import MypagePage from '../../src/components/units/mypage/Mypage.container'
+import MypagePage from '../../src/components/units/mypage/main/Mypage.container'
 
 export default function Mypage():JSX.Element{
     return <MypagePage/>
