+++ alcoholmap/src/frontend/src/components/units/mypage/main/Mypage.styles.tsx
