+++ alcoholmap/src/frontend/src/components/units/mypage/main/Mypage.types.ts
