+++ alcoholmap/src/frontend/src/components/units/mypage/wishlist/WishListSlider.styles.tsx
